Use iframe to implement common page header
We could also accomplish this with server-side includes. One advantage
of this is that the client can implement the iframe trick without
depending on the server to do the work. That means that the site will
function on local Windows machines (just from local files) or on the
GitHub.io site.

A disadvantage of iframes is that you must specify the iframe size in
the iframe element, and cannot let it layout with the natural size of
the iframe content itself. If there are other issues we find later, we
can go with server-side include (SSI) later.

For a test illustrating SSI, checkout this commit and refer to the files
ssi-test.shtml and ssi-test-frag.html.